Break out AttributesDecoderTests
