+++ Tests/MusicXMLTests/DecoderTests.swift
@@ -11,34 +11,7 @@ import MusicXML
 
 class DecoderTests: XCTestCase {
 
-    func testTime() {
-        let xml = """
-        <time>
-            <beats>4</beats>
-            <beat-type>4</beat-type>
-        </time>
-        """
-        let _ = try! XMLDecoder().decode(MusicXML.Attributes.Time.self, from: xml.data(using: .utf8)!)
-    }
-
-    func testKey() {
-        let xml = """
-        <key>
-            <fifths>0</fifths>
-        </key>
-        """
-        let _ = try! XMLDecoder().decode(MusicXML.Attributes.Key.self, from: xml.data(using: .utf8)!)
-    }
-
-    func testClef() {
-        let xml = """
-        <clef>
-            <sign>G</sign>
-            <line>2</line>
-        </clef>
-        """
-        let _ = try! XMLDecoder().decode(MusicXML.Attributes.Clef.self, from: xml.data(using:. utf8)!)
-    }
+    
 
     func testMeasure() {
         let xml = """
